+++ backend/src/main/java/com/cpweb/backend/repositories/RecetteRepository.java
@@ -1,8 +1,7 @@
 package com.cpweb.backend.repositories;
 
-import com.cpweb.backend.models.Categorie;
 import com.cpweb.backend.models.Recette;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RecetteRepository extends JpaRepository<Recette, Integer> {
+public interface RecetteRepository extends JpaRepository<Recette, Long> {
 }
